HierAMP/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
biopython>=1.81
tqdm>=4.65.0
wandb>=0.15.0
einops>=0.7.0
scipy>=1.11.0
PyYAML>=6.0
tensorboard>=2.14.0